+++ tests/screenshot/apps/components/WizardPanels.js
@@ -67,6 +67,33 @@ const WizardPanelTests = withConfig({
 			</WizardPanels>
 		)
 	},
+	// Test no buttons
+	<WizardPanels index={0}>
+		<Panel nextButton={false} title="First Panel Title">View 1</Panel>
+		<Panel prevButton={false} title="Second Panel Title">View 2</Panel>
+	</WizardPanels>,
+	<WizardPanels index={1}>
+		<Panel nextButton={false} title="First Panel Title">View 1</Panel>
+		<Panel prevButton={false} title="Second Panel Title">View 2</Panel>
+	</WizardPanels>,
+	{
+		locale: 'ar-SA',
+		component: (
+			<WizardPanels index={0}>
+				<Panel nextButton={false} title="First Panel Title">View 1</Panel>
+				<Panel prevButton={false} title="Second Panel Title">View 2</Panel>
+			</WizardPanels>
+		)
+	},
+	{
+		locale: 'ar-SA',
+		component: (
+			<WizardPanels index={1}>
+				<Panel nextButton={false} title="First Panel Title">View 1</Panel>
+				<Panel prevButton={false} title="Second Panel Title">View 2</Panel>
+			</WizardPanels>
+		)
+	},
 	// Test custom buttons
 	<WizardPanels index={1} title="WizardPanel">
 		<Panel>View 1</Panel>